+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Improve attribution controls accessibility. See [#359](https://github.com/maplibre/maplibre-gl-js/issues/359)
 - Allow maxPitch value up to 85, use values greater than 60 at your own risk (#574)
 - `getImage` uses createImageBitmap when supported (#650)
+- Allow setting a custom pixel ratio by adding a `MapOptions#pixelRatio` property and a `Map#setPixelRatio` method (#769)
 - *...Add new stuff here...*
 
 ### 🐞 Bug fixes